Fix cyclepainting command

+++ common/src/main/java/com/ptsmods/morecommands/commands/unelevated/CyclePaintingCommand.java
@@ -37,8 +37,9 @@ public class CyclePaintingCommand extends Command {
         }
 
         ResourceLocation oldArt = Compat.get().getPaintingVariant(painting);
-
         Compat.get().setPaintingVariant(painting, variant == null ? Compat.get().nextPaintingVariant(oldArt) : variant);
+
+        if (Version.getCurrent().isOlderThan(Version.V1_19)) return 1; // Below code only necessary on old versions
         BlockPos pos = Compat.get().blockPosition(painting);
         Entity painting0 = MoreCommands.cloneEntity(painting, false);
         painting.kill();
